requirements.txt
uvicorn
pywebview
fastapi
requests
sqlmodel